🔒 chore(pyproject.toml): update python and pytest-httpx dependencies (#9)
* 🔒 chore(pyproject.toml): update python and pytest-httpx dependencies
This commit updates the python dependency to version 3.7, which is the minimum version required by other dependencies. The pytest-httpx dependency is also updated to version 0.22.0, which is the latest version available at the time of this commit.

* 🚀 chore(pyproject.toml): bump version to 1.0.2
This commit updates the version of the project to 1.0.2 in the pyproject.toml file. This is a chore commit as it does not add any new features or fix any bugs, but it is important to keep the version up to date to reflect the changes made to the project.

* 🔧 chore(tests.yml): update python-version matrix to include 3.10 and 3.11
    The python-version matrix has been updated to include the latest versions of Python, 3.10 and 3.11. This ensures that the application is tested against the latest versions of Python and helps to identify any compatibility issues early on.