🚀 chore(curlify2): refactor to_curl function to a class method (#14)
* 🔥 chore(pyproject.toml): remove responses from dev dependencies
The responses package is no longer needed in the project, so it has been removed from the dev dependencies.

* 🔧 chore(test_to_curl.py): refactor test_returns_curl_string_using_requests_module and test_returns_curl_string_using_httpx_module
The tests have been refactored to use the response object returned by the to_curl function instead of assigning it to a variable and then comparing it to a string. This improves the readability of the tests and makes it easier to identify the source of any failures.

* 🚀 chore(.gitignore): add .venv and dist directories to .gitignore
🚀 chore(curlify2): refactor to_curl function to a class method
🚀 chore(pyproject.toml): bump version to 2.0.0
The .venv and dist directories are now ignored by git. The to_curl function in curlify2 is now a class method, which improves readability and maintainability. The version in pyproject.toml is now 2.0.0 to reflect the changes made.

* 📦 chore(pyproject.toml): add responses package
The responses package is added to the project dependencies in the pyproject.toml file. This package is used for mocking HTTP requests in tests.

* 🐛 fix(curlify.py): fix body decoding
✨ feat(app.py): add endpoint to convert http request to curl command
The body decoding was fixed in curlify.py. The method now returns the decoded body if it is a bytes object. In app.py, a new endpoint was added to convert an HTTP request to a curl command. The endpoint receives a URL via a POST request, sends a GET request to the URL, and converts the request to a curl command using curlify. The command is then returned to the user.

* 🎨 style(README.md): refactor code to use Curlify class instead of to_curl function
The code has been refactored to use the Curlify class instead of the to_curl function from the curlify2 module. This improves readability and makes the code more concise.